Migrate KTX transitive text data out of MakeReleaseNotesTask
